Use gamma in filter design and combine Lanczos and Gaussian
Introduces a gamma of 0.43 in the design of the Gaussian filter, to estimate
the human eye's response to brightness levels. Set the cut-off based on half
perceived brightness.

Also, create a new filter that is an average of the Gaussian and Lanczos3
filter. This balances the pros and cons of each and seems to be a good choice
in many cases.